Generator: exit placement ongoing
? Have most of it done now, but realized you can't let them land in a 
potential exit, notably if they can move horizontally and hit another 
vertical exit it's janky. So you have to jump an extra tile into the 
next room to avoid this, but THAT means that you might end up literally 
in some water, so I need to give that some thought. Boo.