Added five more waters
? Relatively smooth sailing with the process for now. It's generally 
possible to find the correct tiles in the source by looking for animated 
tiles specifically, as most water is animated. Still can involve a bit 
of hit and miss and sometimes I've hand-coded them, but by and large it 
works.

? A few of the waters aren't really working out in some games. Notably 
those that have hyper specific "topographies"? As in they represent a 
really specific stream or eddying water around rocks or something. 
Really the water I'm displaying here needs to roughly resemble an open 
body of water to a) make sense relative to the original game and b) be 
non-exhausting to put together.